pyproject.toml
[tool.poetry]
name = "ai-on-telegram"
version = "0.1.0"
description = ""
authors = ["Nick <nickphilomath@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pytelegrambotapi = "^4.25.0"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
